requirements.txt
Django==4.2.12
djangorestframework==3.15.1
drf-yasg==1.21.7
django-filter==24.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
channels==4.1.0
channels-redis==4.2.0
python-dotenv==1.0.1
requests==2.31.0
